# bu-analytics-gen-ai-midas/testing/generated/openapi_from_inventory.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MIDAS route inventory stub",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://placeholder.invalid",
      "description": "override via case.call(base_url=\u2026)"
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/api/v1/auth/users": {
      "get": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/api/v1/auth/users/{user_id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      },
      "put": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/api/v1/auth/verify-token": {
      "post": {
        "responses": {
          "200": {
            "description": "inventory stub \u2014 validate transport only"
          },
          "401": {
            "description": "unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
...
  }
}